requirements.txt
fastapi
uvicorn
pdf2image
pytesseract
opencv-python-headless
Pillow
spacy
scikit-learn
